refactor: move cleanup function to separated file
