+++ system-test/reader_client_test.ts
@@ -24,6 +24,7 @@ import {ClientOptions} from 'google-gax';
 import * as customerRecordProtoJson from '../samples/customer_record.json';
 import * as bigquerystorage from '../src';
 import * as reader from '../src/reader';
+import {cleanupDatasets} from './util';
 import {RecordBatch, Table, tableFromIPC} from 'apache-arrow';
 
 type ReadRowsResponse =
@@ -103,7 +104,7 @@ describe('reader.ReaderClient', () => {
   };
 
   before(async () => {
-    await deleteDatasets();
+    await cleanupDatasets(bigquery, GCLOUD_TESTS_PREFIX);
 
     await bigquery.createDataset(datasetId);
   });
@@ -687,47 +688,4 @@ describe('reader.ReaderClient', () => {
       }
     });
   });
-
-  // Only delete a resource if it is older than 24 hours. That will prevent
-  // collisions with parallel CI test runs.
-  function isResourceStale(creationTime: number) {
-    const oneDayMs = 86400000;
-    const now = new Date();
-    const created = new Date(creationTime);
-    return now.getTime() - created.getTime() >= oneDayMs;
-  }
-
-  async function deleteDatasets() {
-    let [datasets] = await bigquery.getDatasets().catch(e => {
-      console.log('bigquery.getDatasets() failed');
-      console.log(e);
-      return [[]];
-    });
-    datasets = datasets.filter(dataset =>
-      dataset.id?.includes(GCLOUD_TESTS_PREFIX),
-    );
-
-    for (const dataset of datasets) {
-      const isDatasetStable = await dataset
-        .getMetadata()
-        .then(res => {
-          const [metadata] = res;
-          const creationTime = Number(metadata.creationTime);
-          return isResourceStale(creationTime);
-        })
-        .catch(e => {
-          console.log(`dataset(${dataset.id}).getMetadata() failed`);
-          console.log(e);
-          return false;
-        });
-      if (isDatasetStable) {
-        try {
-          await dataset.delete({force: true});
-        } catch (e) {
-          console.log(`dataset(${dataset.id}).delete() failed`);
-          console.log(e);
-        }
-      }
-    }
-  }
 });
